feat: share tailwind config
