+++ apps/deploy-web/package.json
@@ -116,8 +116,6 @@
     "sharp": "^0.30.3",
     "stripe": "^10.14.0",
     "tailwind-merge": "^2.0.0",
-    "tailwind-scrollbar": "^3.0.5",
-    "tailwindcss-animate": "^1.0.7",
     "tss-react": "^4.8.5",
     "use-state-with-callback": "^3.0.2",
     "usehooks-ts": "^2.9.1",
@@ -130,7 +128,7 @@
     "@akashnetwork/dev-config": "*",
     "@keplr-wallet/types": "^0.10.15",
     "@next/bundle-analyzer": "^14.0.1",
-    "@tailwindcss/typography": "^0.5.12",
+    "@akashnetwork/tailwind-config": "*",
     "@types/auth0": "^2.35.3",
     "@types/file-saver": "^2.0.5",
     "@types/js-yaml": "^4.0.5",
