+++ apps/deploy-web/postcss.config.js
@@ -1,8 +1,1 @@
-module.exports = {
-  plugins: {
-    "postcss-import": {},
-    "tailwindcss/nesting": "postcss-nesting",
-    tailwindcss: {},
-    autoprefixer: {}
-  }
-};
+module.exports = require('../../packages/tailwind-config/postcss.config')
