+++ apps/stats-web/package.json
@@ -50,12 +50,12 @@
     "react-modern-drawer": "^1.2.2",
     "react-query": "^3.39.3",
     "tailwind-merge": "^2.0.0",
-    "tailwindcss-animate": "^1.0.7",
     "usehooks-ts": "^2.9.1",
     "vaul": "^0.8.0",
     "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@akashnetwork/tailwind-config": "*",
     "@types/json2csv": "^5.0.7",
     "@types/node": "20.14.0",
     "@types/react": "18.2.0",
