+++ packages/tailwind-config/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@akashnetwork/tailwind-config",
+  "version": "1.0.0",
+  "files": [
+    "tailwind.config.ts"
+  ],
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "devDependencies": {
+    "@tailwindcss/typography": "0.5.13",
+    "prettier-plugin-tailwindcss": "0.6.1",
+    "tailwind-scrollbar": "3.1.0",
+    "tailwindcss": "3.4.4",
+    "tailwindcss-animate": "1.0.7"
+  }
+}
