+++ packages/ui/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@akashnetwork/ui",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": ""
+}
