+++ packages/ui/postcss.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../tailwind-config/postcss.config')
