+++ packages/ui/tailwind.config.ts
@@ -1,0 +1,1 @@
+module.exports = require('../tailwind-config/tailwind.config')
